refactor(cardwired): split PCI and display monitors
