+++ crates/cardwire-daemon/src/tasks/mod.rs
@@ -1,8 +1,10 @@
 mod battery_switch;
 mod fdo_list;
+mod monitor_display;
 mod monitor_udev;
 mod watch_power_state;
 
 pub use battery_switch::watch_battery_status;
-pub use monitor_udev::monitor_hardware_changes;
+pub use monitor_display::monitor_display_changes;
+pub use monitor_udev::monitor_pci_changes;
 pub use watch_power_state::watch_power_state;
